Добавлена аутентификация и авторизация на основе ресурсов и ролей

+++ src/Application/AdBoard.Application.AppData/Contexts/User/Services/UserService.cs
@@ -1,7 +1,11 @@
-﻿using AdBoard.Application.AppData.Contexts.Advert.Repositories;
+﻿using System.Security.Claims;
+using AdBoard.Application.AppData.Authorization.Requirements.Operation;
+using AdBoard.Application.AppData.Contexts.Advert.Repositories;
 using AdBoard.Application.AppData.Contexts.Advert.Services;
 using AdBoard.Contracts.User;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 
 namespace AdBoard.Application.AppData.Contexts.User.Services;
 
@@ -11,12 +15,20 @@ using User = AdBoard.Domain.User.User;
 public class UserService : IUserService
 {
     private readonly IUserRepository _repository;
+    private readonly IHttpContextAccessor _httpContextAccessor;
     private readonly IMapper _mapper;
+    private readonly IAuthorizationService _authorizationService;
 
-    public UserService(IUserRepository repository, IMapper mapper)
+    public UserService(
+        IUserRepository repository, 
+        IHttpContextAccessor httpContextAccessor, 
+        IMapper mapper, 
+        IAuthorizationService authorizationService)
     {
         _repository = repository;
+        _httpContextAccessor = httpContextAccessor;
         _mapper = mapper;
+        _authorizationService = authorizationService;
     }
 
     public Task<ShortUserDto[]> GetAll(CancellationToken cancellationToken)
@@ -29,6 +41,20 @@ public class UserService : IUserService
         return _repository.GetById(id, cancellationToken);
     }
 
+    public async Task<UserDto> GetCurrent(CancellationToken cancellationToken)
+    {
+            var claims = _httpContextAccessor.HttpContext?.User.Claims;
+            if (claims is null) throw new ArgumentNullException(nameof(claims));
+
+            var idString = claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (idString is null) throw new ArgumentNullException(nameof(idString));
+
+            var id = Guid.Parse(idString);
+            var user = await _repository.GetById(id, cancellationToken);
+        
+            return _mapper.Map<UserDto>(user);
+        }
+
     public async Task<ShortUserDto> Add(CreateUserDto dto, CancellationToken cancellationToken)
     {
         var user = _mapper.Map<User>(dto);
@@ -37,12 +63,23 @@ public class UserService : IUserService
 
     public async Task<ShortUserDto> Update(Guid id, UpdateUserDto dto, CancellationToken cancellationToken)
     {
+        var userDto = await _repository.GetById(id, cancellationToken);
+        if (userDto is null) throw new ArgumentException("User with id not exist", nameof(id));
+        
+        var authorizationResult = await _authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext!.User, userDto, OperationsList.Update);
+        if (!authorizationResult.Succeeded) throw new UnauthorizedAccessException("User hasn't permissions to this user");
+        
         var user = _mapper.Map<User>(dto);
+        user.Role = userDto.Role;
         return await _repository.Update(id, user, cancellationToken);
     }
 
     public async Task<bool> Delete(Guid id, CancellationToken cancellationToken)
     {
+        var userDto = await _repository.GetById(id, cancellationToken);
+        var authorizationResult = await _authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext!.User, userDto, OperationsList.Delete);
+        if (!authorizationResult.Succeeded) throw new UnauthorizedAccessException("User hasn't permissions to this user");
+        
         return await _repository.Delete(id, cancellationToken);
     }
 }
